+++ CHANGELOG.md
@@ -20,6 +20,8 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#1702](https://github.com/Pycord-Development/pycord/pull/1702))
 - New user flag: `PublicUserFlags.active_developer`.
   ([#1776](https://github.com/Pycord-Development/pycord/pull/1776))
+- New application flag: `ApplicationFlags.active_application`.
+  ([#1776](https://github.com/Pycord-Development/pycord/pull/1776))
 
 ### Fixed